requirements.txt
mkdocs==1.3.0
mkdocs-material>=8
mkdocs-material-extensions
mkdocs-macros-plugin
GitPython
requests==2.26.0
#mkdocs-with-pdf